backend/requirements.txt
Flask==2.0.3
python-decouple==3.6
gunicorn==20.1.0
requests==2.28.1
flasgger==0.9.5
werkzeug==2.0.3
pymongo==4.3.3
PyJWT==2.6.0
cryptography==41.0.3
Flask-Cors==3.0.10
flask-jwt-extended==4.4.4
SQLAlchemy==1.4.46
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Bcrypt==1.0.1
stripe==5.4.0
celery
redis
python-dotenv==1.0.0
python-docx
openpyxl
PyPDF2
fpdf
Flask-SocketIO
python-jose==3.3.0  # Para manipulação de JWT em compatibilidade com auth0/outros provedores
boto3==1.28.38  # Para integração com AWS S3

# Novas dependências para funcionalidades de eBook
pymupdf==1.21.1  # Para manipulação de PDFs
ebooklib==0.17.1  # Para criação de EPUBs
pillow==9.5.0  # Para manipulação de imagens
openai==1.3.0  # Para integração com OpenAI
stability-sdk==0.8.0  # Para integração com Stability AI (opcional)
# Para a integração com Canva, usamos diretamente a API REST via requests
markdown==3.4.3  # Para conversão de markdown para outros formatos
jinja2==3.1.2  # Para templates
weasyprint==59.0  # Para conversão de HTML para PDF

# Dependências para processamento de fichas médicas
python-docx==0.8.11  # Para processamento de arquivos .docx
pandas==1.5.3  # Para processamento de arquivos Excel
numpy==1.23.5  # Versão compatível com pandas 1.5.3
openpyxl==3.1.2  # Para suporte a arquivos Excel modernos
xlrd==2.0.1  # Para suporte a arquivos Excel antigos
python-dateutil==2.8.2  # Para manipulação de datas

# Novas dependências para melhorar documentação da API e validação
apispec==6.0.2  # Para gerar especificações OpenAPI
flask-marshmallow==0.15.0  # Para serialização e validação
marshmallow-sqlalchemy==0.29.0  # Para integração do marshmallow com SQLAlchemy
marshmallow==3.19.0  # Para serialização e validação de dados
marshmallow-enum==1.5.1  # Para suporte a enums no marshmallow
apispec-webframeworks==0.5.2  # Para integração do apispec com Flask
pyyaml==6.0  # Para lidar com arquivos YAML na documentação

# Dependência para integração do Swagger com APIs RESTful
flask-restful==0.3.10  # Para suportar o Flasgger com APIs RESTful

# Dependências para testes
sseclient==0.0.27  # Para testes de streaming de SSE (Server-Sent Events)
